django_app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
openai>=1.3,<2.0
requests>=2.25,<3.0
pytz>=2023.3
